Fixing '#' issue in Embed Gen
